Inline partial helpers locally per maintainer feedback
Duplicate functools.partial unwrap/bind helpers in draw.py and specs.py instead of a shared pennylane/_functools_partial module, remove the tach.toml entry, and fix Catalyst specs tests to use fixed gate counts without static_argnums.
